booking resolver refactored
