+++ src/graphql/resolvers/booking/index.js
@@ -1,0 +1,9 @@
+import { bookings } from './bookings'
+import { createBookingEvent } from './createBookingEvent'
+import { cancelBooking } from './cancelBooking'  
+
+export const bookingResolver = {
+  ...bookings,
+  ...createBookingEvent,
+  ...cancelBooking
+}
